move `Combobox` for clarity when reading the code
and avoid a missinterpretation as there its implementation is specific to the SelectPlace component
