+++ src/lib/components/SelectPlace/Combobox/ClearInputIcon.svelte
