+++ src/lib/components/SelectPlace/Combobox/index.ts
@@ -1,3 +1,2 @@
 import Combobox from "./Combobox.svelte";
-
-export default Combobox;
+export { Combobox };
